requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
onnxruntime>=1.20.0
rapidocr>=1.4.0
pillow>=10.0.0
python-multipart>=0.0.6
pdf2image>=1.17.0
requests>=2.31.0
